go schemabuilder: Move Schema-specific API code to separate file.
Summary: Moves the Schema type out of the reflect file and into it's own
file, this is a first step to help clean up some of the code and
increase understanding.
